Move GitHub stuff over :D
